requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
bcrypt==3.1.4
certifi==2018.11.29
chardet==3.0.4
configparser==3.5.0
crcmod==1.7
cryptography==2.3
Django==1.11.18
django-debug-toolbar==1.11
dnspython==1.15.0
docutils==0.14
entrypoints==0.2.3.post3
enum34==1.1.6
folder-color==0.0.86
folder-color-common==0.0.86
futures==3.2.0
idna==2.8
iexfinance==0.3.5
ipaddress==1.0.17
isort==4.3.4
keyring==15.1.0
keyrings.alt==3.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector-python==8.0.11
mysql-utilities==1.6.4
nltk==3.3
numpy==1.16.1
oauthlib==2.1.0
pandas==0.24.1
paramiko==2.4.1
pexpect==4.6.0
protobuf==3.0.0
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
PyGObject==3.30.1
pylint==1.9.3
PyNaCl==1.2.1
pyodbc==4.0.22
pysqlite==2.7.0
python-dateutil==2.8.0
pytz==2018.9
pyxdg==0.25
requests==2.21.0
requests-oauthlib==1.0.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.2.4
statistics==1.0.3.5
TA-Lib==0.4.17
trash-cli==0.12.9.14
twython==3.7.0
urllib3==1.24.1
virtualenv==16.2.0
wrapt==1.10.11
